This commit is most probably wrong, but lets store it
The issue is that when the PCB is provided power then the led strip
quickly lights up, then goes of. Afterwards wled turns it on again.

I though this is because the wiring of the relay is wrong. The
documentation shows the default state to be in reverse of what the
downloadbale schematic provides
(https://www.azettler.com/pdfs/az943.pdf).
Then the relay would be on when no power is applied through the coil.

However that does not really match the sounds it makes. During the quick
on time of the led strip there are two click sounds instead of one which
should be the case if the above holds true.
Unless the relay per default clicks once when power is provided even
when the coil is low.